Refactoring tests + removing additional code
