+++ tests/functional/event_handler/_pydantic/test_openapi_validation_middleware.py
@@ -2,7 +2,7 @@ import json
 from dataclasses import dataclass
 from enum import Enum
 from pathlib import PurePath
-from typing import List, Optional, Set, Tuple
+from typing import List, Optional, Tuple
 
 import pytest
 from pydantic import BaseModel
@@ -1173,7 +1173,17 @@ def test_none_returned_for_optional_type(gw_event):
     assert result["body"] == "null"
 
 
-def test_valid_model_returned_for_optional_type(gw_event):
+@pytest.mark.parametrize(
+    "path, body",
+    [
+        ("/empty_dict", {}),
+        ("/empty_list", []),
+        ("/none", "null"),
+        ("/empty_string", ""),
+    ],
+    ids=["empty_dict", "empty_list", "none", "empty_string"],
+)
+def test_none_returned_for_falsy_return(gw_event, path, body):
     # GIVEN an APIGatewayRestResolver with validation enabled
     app = APIGatewayRestResolver(enable_validation=True)
 
@@ -1181,121 +1191,13 @@ def test_valid_model_returned_for_optional_type(gw_event):
         name: str
         age: int
 
-    @app.get("/valid_optional")
-    def handler_valid_optional() -> Optional[Model]:
-        return Model(name="John", age=30)
+    @app.get(path)
+    def handler_none_allowed() -> Model:
+        return body
 
-    # WHEN returning a valid model for an Optional type
-    gw_event["path"] = "/valid_optional"
+    # WHEN returning None for an Optional type
+    gw_event["path"] = path
     result = app(gw_event, {})
 
-    # THEN it should succeed and return the serialized model
-    assert result["statusCode"] == 200
-    assert json.loads(result["body"]) == {"name": "John", "age": 30}
-
-
-def test_serialize_response_without_field(gw_event):
-    # GIVEN an APIGatewayRestResolver with validation enabled
-    app = APIGatewayRestResolver(enable_validation=True)
-
-    # WHEN a handler is defined without return type annotation
-    @app.get("/test")
-    def handler():
-        return {"message": "Hello, World!"}
-
-    gw_event["path"] = "/test"
-
-    # THEN the handler should be invoked and return 200
-    # AND the body must be a JSON object
-    response = app(gw_event, None)
-    assert response["statusCode"] == 200
-    assert response["body"] == '{"message":"Hello, World!"}'
-
-
-def test_serialize_response_list(gw_event):
-    """Test serialization of list responses containing complex types"""
-    # GIVEN an APIGatewayRestResolver with validation enabled
-    app = APIGatewayRestResolver(enable_validation=True)
-
-    # WHEN a handler returns a list containing various types
-    @app.get("/test")
-    def handler():
-        return [{"set": [1, 2, 3]}, {"simple": "value"}]
-
-    gw_event["path"] = "/test"
-
-    # THEN the response should be properly serialized
-    response = app(gw_event, None)
-    assert response["statusCode"] == 200
-    assert response["body"] == '[{"set":[1,2,3]},{"simple":"value"}]'
-
-
-def test_serialize_response_nested_dict(gw_event):
-    """Test serialization of nested dictionary responses"""
-    # GIVEN an APIGatewayRestResolver with validation enabled
-    app = APIGatewayRestResolver(enable_validation=True)
-
-    # WHEN a handler returns a nested dictionary with complex types
-    @app.get("/test")
-    def handler():
-        return {"nested": {"date": "2000-01-01", "set": [1, 2, 3]}, "simple": "value"}
-
-    gw_event["path"] = "/test"
-
-    # THEN the response should be properly serialized
-    response = app(gw_event, None)
-    assert response["statusCode"] == 200
-    assert response["body"] == '{"nested":{"date":"2000-01-01","set":[1,2,3]},"simple":"value"}'
-
-
-@dataclass
-class Person:
-    name: str
-    birth_date: str
-    scores: Set[int]
-
-
-def test_serialize_response_dataclass(gw_event):
-    """Test serialization of dataclass responses"""
-    # GIVEN an APIGatewayRestResolver with validation enabled
-    app = APIGatewayRestResolver(enable_validation=True)
-
-    # WHEN a handler returns a dataclass instance
-    @app.get("/test")
-    def handler():
-        return Person(name="John Doe", birth_date="1990-01-01", scores=[95, 87, 91])
-
-    gw_event["path"] = "/test"
-
-    # THEN the response should be properly serialized
-    response = app(gw_event, None)
-    assert response["statusCode"] == 200
-    assert response["body"] == '{"name":"John Doe","birth_date":"1990-01-01","scores":[95,87,91]}'
-
-
-def test_serialize_response_mixed_types(gw_event):
-    """Test serialization of mixed type responses"""
-    # GIVEN an APIGatewayRestResolver with validation enabled
-    app = APIGatewayRestResolver(enable_validation=True)
-
-    # WHEN a handler returns a response with mixed types
-    @app.get("/test")
-    def handler():
-        person = Person(name="John Doe", birth_date="1990-01-01", scores=[95, 87, 91])
-        return {
-            "person": person,
-            "records": [{"date": "2000-01-01"}, {"set": [1, 2, 3]}],
-            "metadata": {"processed_at": "2050-01-01", "tags": ["tag1", "tag2"]},
-        }
-
-    gw_event["path"] = "/test"
-
-    # THEN the response should be properly serialized
-    response = app(gw_event, None)
-    assert response["statusCode"] == 200
-    expected = {
-        "person": {"name": "John Doe", "birth_date": "1990-01-01", "scores": [95, 87, 91]},
-        "records": [{"date": "2000-01-01"}, {"set": [1, 2, 3]}],
-        "metadata": {"processed_at": "2050-01-01", "tags": ["tag1", "tag2"]},
-    }
-    assert json.loads(response["body"]) == expected
+    # THEN it should succeed
+    assert result["statusCode"] == 422
